D-37449 - display the filtercount in filterComponent
