+++ plugins/dai-release/src/components/HomePageComponent/ReleaseMetaContent.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import {TemplateGitMetaInfo} from "@digital-ai/plugin-dai-release-common";
+
+export const getActiveReleaseMetaContent = (
+    modalPopupData: TemplateGitMetaInfo,
+) => {
+    return (
+        <div>
+            <div className="modal-body version-control">
+                {!modalPopupData?.name ? (
+                    <p>No data available.</p>
+                ) : (
+                    <div className="" />
+                )}
+            </div>
+        </div>
+    );
+};
