+++ plugins/dai-release/src/components/TemplateHomePageComponent/TemplateMetaContent.tsx
@@ -52,19 +52,3 @@ export const getTemplateMetaContent = (
     </div>
   );
 };
-
-export const getActiveReleaseMetaContent = (
-  modalPopupData: TemplateGitMetaInfo,
-) => {
-  return (
-    <div>
-      <div className="modal-body version-control">
-        {!modalPopupData?.name ? (
-          <p>No data available.</p>
-        ) : (
-          <div className="" />
-        )}
-      </div>
-    </div>
-  );
-};
